+++ components/Editor.tsx
@@ -19,7 +19,7 @@ import { Toggle } from "./ui/toggle";
 import { Input } from "./ui/input";
 import { useToast } from "./hooks/use-toast";
 import { Toaster } from "./ui/toaster";
-import CopyLinkButton from "./DrawingLink";
+import CopyLinkButton from "./LinkButton";
 
 function ColorSelectorToggleButton({
   tooltip,
